Split useFilter so the local filtering part can be omitted
